Add experimental ability to separate queries into segments.
Code attempts to match the segments of a select query (select/from/where/etc) and include these individual parts in the result data.
Known to be buggy/imperfect - hence experimental status - a more robust implementation could be done with an SQL parser, assuming one could be configured to understand/ignore the CFML tags which occur in raw queries.