Workaround FW/1 flaw, to still work when directory name contains dot.
FW/1's controller discovery fails when the directory has a dot in it, as
is the case with GitHub's default download. Setting cfcbase to blank
should fix this, but the cfcFilePath function assumes blank means webroot
(instead of current directory), so that function is overriden to change
this.

Fixes #11 on Github.